The current status of the QLogic target driver in the SCST project is as follows: - Building the QLogic target driver is possible by following the instructions  in qla2x00t/qla2x00-target/README. - Some but not all of the qla2x00-target code is included in the output of  scripts/generate-kernel-patch. This is unfortunate because it is the output  of this script that is posted to the LKML each time SCST is submitted for  review. - Most of the qla2x00-target code is checked during the nightly build by  checkpatch. None of the qla2x00-target code is checked during the nightly  build by sparse.
The patch below makes sure that all of the qla2x00-target code is included in
the generate-kernel-patch output and also that all of the qla2x00-target code
is checked during the nightly build by both checkpatch and sparse. This has
been realized by including the diff of the linux-2.6.26.8/drivers/scsi/qla2xxx
source files and the source files in qla2x00t/*[ch] as individual patches under
qla2x00t/in-tree-patches. These patches are generated from downloaded
linux-2.6.26.8 sources and the sources in the SCST SVN repository.

The next step is to fix the checkpatch and sparse warnings triggered by the
qla2x00-target code and to port the latest qla2xxx patches to the SCST
qla2x00t implementation.

Signed-off-by: Bart Van Assche <bart.vanassche@gmail.com>