Eliminated msleep(3000) in scst_sysfs_cleanup().
Allocated scst_sysfs_root_kobj dynamically instead of statically because I found the following in Documentation/kobject.txt and for no other reason:

Because kobjects are dynamic, they must not be declared statically or on
the stack, but instead, always allocated dynamically.  Future versions of
the kernel will contain a run-time check for kobjects that are created
statically and will warn the developer of this improper usage.