scst: Fix debug_print_with_prefix()
Terminate lines passed to debug_print_with_prefix() with a newline. Make
sure that a formatting specification in an argument past "fmt" is printed
literally instead of being interpreted as a format specification.

The newlines are only missing when sending kernel logging to a remote
system via netconsole and not when using local logging

Signed-off-by: Bart Van Assche <bvanassche@acm.org>