scst_vdisk: Remove a superfluous cast
The 'iov_base' member of 'struct iovec' is declared in
<uapi/linux/uio.h> with type 'void __user *'. Adding an integer
to a void pointer has the same effect as adding an integer to
a char pointer in the C language. This means that it is safe to
remove the cast in front of the eiv->iov_base pointer.

Please note that a similar construct already exists in the iSCSI
target driver. From the nthread.c source file:

			iop->iov_base += rest;
			iop->iov_len -= rest;

This patch does not change any functionality.

Signed-off-by: Bart Van Assche <bvanassche@acm.org>