ib_srpt: Change default behavior from using SRQ to not using SRQ
Although the non-SRQ mode needs more resources that mode has the
advantage that the end-to-end flow control mechanism is used.

From the IB spec:

C9-150.2.1: For QPs that are not associated with an SRQ, each HCA
receive queue shall generate end-to-end flow control credits. If
a QP is associated with an SRQ, the HCA receive queue shall not
generate end-to-end flow control credits.