scst_vdisk: Make the filename modifiable
When VAAI support is not needed can be convenient to use the
vdisk_blockio handler to handle I/O redirection between nodes in a
H.A. setup. Making the filename modifiable allows to keep the LUN
associations of an SCST device that is used for I/O redirection in
such a setup. This is a slightly modified version of a patch
provided by Marc Smith <marc.smith@parodyne.com>.