qla2x00t-32gbit: Fix race conditions in the abort handling code
In the *_done() functions, instead of returning early if sp->ref_count >= 2,
only decrement sp->ref_count. In qla2xxx_eh_abort(), instead of deciding
what to do based on the value of sp->ref_count, decide which action to take
depending on the completion status of the firmware abort.