qla_isp: Remove this driver
This driver is obsolete and is no longer supported since a long time, so
remove it. From the README of this driver (this note was added in 2009):
"From other side, qla2x00t is simpler, smaller and much better tested
on 22xx and 23xx, hence perform more reliable and, thus, is recommended
for these adapters. Since 24xx/25xx become fully supported on qla2x00t
we encourage users to switch to this driver."